+++ changelog.md
@@ -1,3 +1,6 @@
+## [unreleased] - 2019-06-14
+- Add example for basic mongo setup
+
 ## [1.1.0] - 2019-06-05
 - Add redirectUri to configure custom redirect uri's (callback) on service strategies